Conflicts:
	doc/lua_api.txt
	src/client.h
	src/content_sao.cpp
	src/environment.cpp
	src/environment.h
	src/keycode.cpp
	src/map.cpp
	src/network/clientpackethandler.cpp
	src/particles.cpp
	src/particles.h
	src/script/lua_api/l_particles.cpp
	src/server.cpp
	src/server.h
	src/serverobject.cpp
	src/serverobject.h
	src/threading/thread.h

+++ src/particles.h
@@ -119,7 +119,7 @@ class ParticleSpawner
 		float minsize, float maxsize,
 		bool collisiondetection,
 		bool collision_removal,
-		s32 attached_id,
+		u16 attached_id,
 		bool vertical,
 		video::ITexture *texture,
 		u32 id,
@@ -155,7 +155,7 @@ class ParticleSpawner
 	bool m_collisiondetection;
 	bool m_collision_removal;
 	bool m_vertical;
-	s32 m_attached_id;
+	u16 m_attached_id;
 };
 
 /**
